# coding-openapi.yaml
openapi: "3.1.0"
info:
  title: "GPT-API Coding Agent API"
  version: "1.0.0"
  description: "Narrow Custom GPT Coding Agent API with strict state-machine contracts."
paths:
  /repo/overview:
    post:
      summary: "Repo Overview"
      operationId: "repo_overview_repo_overview_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/search:
    post:
      summary: "Repo Search"
      operationId: "repo_search_repo_search_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoSearchRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/read-context:
    post:
      summary: "Repo Read Context"
      operationId: "repo_read_context_repo_read_context_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoReadContextRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/symbols:
    post:
      summary: "Repo Symbols"
      operationId: "repo_symbols_repo_symbols_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoSymbolsRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/instructions:
    post:
      summary: "Repo Instructions"
      operationId: "repo_instructions_repo_instructions_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/dependency-graph:
    post:
      summary: "Repo Dependency Graph"
      operationId: "repo_dependency_graph_repo_dependency_graph_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/test-map:
    post:
      summary: "Repo Test Map"
      operationId: "repo_test_map_repo_test_map_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/relevant-context:
    post:
      summary: "Repo Relevant Context"
      operationId: "repo_relevant_context_repo_relevant_context_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoRelevantContextRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/callgraph:
    post:
      summary: "Repo Callgraph"
      operationId: "repo_callgraph_repo_callgraph_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/references:
    post:
      summary: "Repo References"
      operationId: "repo_references_repo_references_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoReferencesRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/symbol-references:
    post:
      summary: "Repo Symbol References"
      operationId: "repo_symbol_references_repo_symbol_references_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoSymbolReferencesRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/route-map:
    post:
      summary: "Repo Route Map"
      operationId: "repo_route_map_repo_route_map_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoOverviewRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/changed-context:
    post:
      summary: "Repo Changed Context"
      operationId: "repo_changed_context_repo_changed_context_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoChangedContextRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /repo/recent-history-context:
    post:
      summary: "Repo Recent History Context"
      operationId: "repo_recent_history_context_repo_recent_history_context_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepoRecentHistoryRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    RepoChangedContextRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        base_ref:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Base Ref"
      type: "object"
      required:
        - "repo_path"
      title: "RepoChangedContextRequest"
    RepoOverviewRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        max_depth:
          type: "integer"
          title: "Max Depth"
          default: 4
      type: "object"
      required:
        - "repo_path"
      title: "RepoOverviewRequest"
    RepoReadContextRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        files:
          items:
            type: "string"
          type: "array"
          title: "Files"
        max_bytes_per_file:
          type: "integer"
          title: "Max Bytes Per File"
          default: 50000
      type: "object"
      required:
        - "repo_path"
        - "files"
      title: "RepoReadContextRequest"
    RepoRecentHistoryRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        query:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Query"
        max_commits:
          type: "integer"
          title: "Max Commits"
          default: 20
      type: "object"
      required:
        - "repo_path"
      title: "RepoRecentHistoryRequest"
    RepoReferencesRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        symbol:
          type: "string"
          title: "Symbol"
        max_results:
          type: "integer"
          title: "Max Results"
          default: 100
      type: "object"
      required:
        - "repo_path"
        - "symbol"
      title: "RepoReferencesRequest"
    RepoRelevantContextRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        task:
          type: "string"
          title: "Task"
        diagnostics:
          anyOf:
            -
              items:
                additionalProperties: true
                type: "object"
              type: "array"
            -
              type: "null"
          title: "Diagnostics"
        max_files:
          type: "integer"
          title: "Max Files"
          default: 12
      type: "object"
      required:
        - "repo_path"
        - "task"
      title: "RepoRelevantContextRequest"
    RepoSearchRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        query:
          type: "string"
          title: "Query"
        globs:
          anyOf:
            -
              items:
                type: "string"
              type: "array"
            -
              type: "null"
          title: "Globs"
        max_results:
          type: "integer"
          title: "Max Results"
          default: 50
      type: "object"
      required:
        - "repo_path"
        - "query"
      title: "RepoSearchRequest"
    RepoSymbolReferencesRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        symbols:
          items:
            type: "string"
          type: "array"
          title: "Symbols"
        max_results_per_symbol:
          type: "integer"
          title: "Max Results Per Symbol"
          default: 50
      type: "object"
      required:
        - "repo_path"
        - "symbols"
      title: "RepoSymbolReferencesRequest"
    RepoSymbolsRequest:
      properties:
        repo_path:
          type: "string"
          title: "Repo Path"
        files:
          anyOf:
            -
              items:
                type: "string"
              type: "array"
            -
              type: "null"
          title: "Files"
      type: "object"
      required:
        - "repo_path"
      title: "RepoSymbolsRequest"
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
          type: "array"
          title: "Location"
        msg:
          type: "string"
          title: "Message"
        type:
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
...
  securitySchemes:
    ApiKeyAuth:
      type: "apiKey"
      in: "header"
      name: "x-api-key"
servers:
  -
    url: "https://gpt-api.ngrok.app"
security:
  -
    ApiKeyAuth: